requirements.txt
# Core dependencies
openai>=1.0.0              # For OpenAI and DeepSeek LLM calls
python-dotenv>=1.0.0       # For loading environment variables

# FastAPI dependencies
fastapi>=0.104.0           # FastAPI framework
uvicorn[standard]>=0.24.0  # ASGI server for FastAPI
pydantic>=2.0.0           # Data validation

# Optional dependencies (uncomment if needed)
# google-generativeai>=0.3.0  # For Google Gemini support
# duckduckgo-search>=3.8.0   # For DuckDuckGo search (no API key required)
# requests>=2.28.0           # For web search APIs (Serper, Tavily, Brave, Bocha)
